% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compress_file}
\alias{compress_file}
\title{Compress a file using htslib's bgzf format.}
\usage{
compress_file(in_file, compress)
}
\arguments{
\item{in_file}{Input file name.}

\item{compress}{Integer indicating compression level.}
}
\description{
Compress a file using htslib's bgzf format.
}
